refactor(game): decompose package into runtime and loop modules
