+++ tests/test_end_note_view.py
@@ -7,7 +7,9 @@ from types import SimpleNamespace
 
 import pyray as rl
 
-from crimson.game import EndNoteView, GameState, PANEL_TIMELINE_START_MS
+from crimson.frontend.panels.base import PANEL_TIMELINE_START_MS
+from crimson.game.quest_views import EndNoteView
+from crimson.game.types import GameState
 from crimson.persistence import save_status
 from grim.config import ensure_crimson_cfg
 from grim.console import create_console
@@ -43,10 +45,10 @@ def test_end_note_escape_waits_for_close_transition(monkeypatch, tmp_path: Path)
     def _play_sfx(_audio, key, *, rng=None, allow_variants=True) -> None:  # noqa: ARG001
         played.append(key)
 
-    monkeypatch.setattr("crimson.game.update_audio", lambda _audio, _dt: None)
-    monkeypatch.setattr("crimson.game._ensure_texture_cache", lambda _state: _DummyCache())
-    monkeypatch.setattr("crimson.game.play_sfx", _play_sfx)
-    monkeypatch.setattr("crimson.game.rl.is_key_pressed", lambda _key: False)
+    monkeypatch.setattr("crimson.game.quest_views.update_audio", lambda _audio, _dt: None)
+    monkeypatch.setattr("crimson.game.quest_views._ensure_texture_cache", lambda _state: _DummyCache())
+    monkeypatch.setattr("crimson.game.quest_views.play_sfx", _play_sfx)
+    monkeypatch.setattr("crimson.game.quest_views.rl.is_key_pressed", lambda _key: False)
 
     view = EndNoteView(state)
     view.open()
@@ -54,13 +56,16 @@ def test_end_note_escape_waits_for_close_transition(monkeypatch, tmp_path: Path)
     view.update(0.1)
     view.update(0.1)
 
-    monkeypatch.setattr("crimson.game.rl.is_key_pressed", lambda key: int(key) == int(rl.KeyboardKey.KEY_ESCAPE))
+    monkeypatch.setattr(
+        "crimson.game.quest_views.rl.is_key_pressed",
+        lambda key: int(key) == int(rl.KeyboardKey.KEY_ESCAPE),
+    )
     view.update(0.1)
 
     assert played == ["sfx_ui_buttonclick"]
     assert view.take_action() is None
 
-    monkeypatch.setattr("crimson.game.rl.is_key_pressed", lambda _key: False)
+    monkeypatch.setattr("crimson.game.quest_views.rl.is_key_pressed", lambda _key: False)
     action = None
     for _ in range(30):
         view.update(1.0 / 60.0)
@@ -81,10 +86,10 @@ def test_end_note_draw_fades_pause_background_during_close(monkeypatch, tmp_path
         def get_or_load(self, *_args, **_kwargs):  # noqa: ANN001
             return SimpleNamespace(texture=None)
 
-    monkeypatch.setattr("crimson.game.update_audio", lambda _audio, _dt: None)
-    monkeypatch.setattr("crimson.game._ensure_texture_cache", lambda _state: _DummyCache())
-    monkeypatch.setattr("crimson.game.rl.clear_background", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("crimson.game._draw_screen_fade", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("crimson.game.quest_views.update_audio", lambda _audio, _dt: None)
+    monkeypatch.setattr("crimson.game.quest_views._ensure_texture_cache", lambda _state: _DummyCache())
+    monkeypatch.setattr("crimson.game.quest_views.rl.clear_background", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("crimson.game.quest_views._draw_screen_fade", lambda *_args, **_kwargs: None)
 
     view = EndNoteView(state)
     view.open()
